♻️ refactor: type 코드 이동
