+++ email-worker/src/types/nodemailer.ts
@@ -1,0 +1,6 @@
+export interface NodeMailerError extends Error {
+  code?: string;
+  command?: string;
+  response?: string;
+  responseCode?: number;
+}
